Minor Tweaks
Tidied up file structure and front end
