+++ frontend/.gitignore
@@ -13,10 +13,13 @@
 
 # misc
 .DS_Store
+.env
 .env.local
 .env.development.local
 .env.test.local
 .env.production.local
+.ai
+.psd
 
 npm-debug.log*
 yarn-debug.log*
